Fix missing fi in build_integration

+++ build_integration.sh
@@ -65,16 +65,17 @@ pushd servers
             fi
           fi
           # Retry download again
-          if [ ! -f "../$kafka/kafka-bin/bin/kafka-run-class.sh" ]; then
-          echo "Downloading kafka ${kafka} tarball"
-          if hash curl 2>/dev/null; then
-            curl -f https://archive.apache.org/dist/kafka/$kafka/${KAFKA_ARTIFACT}.tgz -o ${KAFKA_ARTIFACT}.tar.gz || curl -f https://archive.apache.org/dist/kafka/$kafka/${KAFKA_ARTIFACT}.tar.gz -o ${KAFKA_ARTIFACT}.tar.gz
-          else
-            echo "curl not found... using wget"
-            if [ -f "${KAFKA_ARTIFACT}.tar.gz" ]; then
-              echo "Using cached artifact: ${KAFKA_ARTIFACT}.tar.gz"
+          if [ ! -f "${KAFKA_ARTIFACT}.tar.gz" && ! -f "${KAFKA_ARTIFACT}.tgz"]; then
+            echo "Downloading kafka ${kafka} tarball"
+            if hash curl 2>/dev/null; then
+              curl -f https://archive.apache.org/dist/kafka/$kafka/${KAFKA_ARTIFACT}.tgz -o ${KAFKA_ARTIFACT}.tar.gz || curl -f https://archive.apache.org/dist/kafka/$kafka/${KAFKA_ARTIFACT}.tar.gz -o ${KAFKA_ARTIFACT}.tar.gz
             else
-              wget -N https://archive.apache.org/dist/kafka/$kafka/${KAFKA_ARTIFACT}.tgz || wget -N https://archive.apache.org/dist/kafka/$kafka/${KAFKA_ARTIFACT}.tar.gz
+              echo "curl not found... using wget"
+              if [ -f "${KAFKA_ARTIFACT}.tar.gz" ]; then
+                echo "Using cached artifact: ${KAFKA_ARTIFACT}.tar.gz"
+              else
+                wget -N https://archive.apache.org/dist/kafka/$kafka/${KAFKA_ARTIFACT}.tgz || wget -N https://archive.apache.org/dist/kafka/$kafka/${KAFKA_ARTIFACT}.tar.gz
+              fi
             fi
           fi
           echo
